refactor(MovieList): 패키지 변경
