+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.movielist.MovieListFragment">
+        tools:context=".presentation.movielist.movie.MovieListFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerMainMovie"
